400 Bad Request - U+2049
Propose EXCLAMATION QUESTION MARK (U+2049) for status 400 Bad request.

`The Web server (running the Web site) thinks that the data stream sent by the client (e.g. your Web browser or our CheckUpDown robot) was 'malformed' i.e. did not respect the HTTP protocol completely. So the Web server was unable to understand the request and process it.``

The `?!` emoji represents confusion that the server experiences when client sends garbage to it.